Refactor to work when before babel-plugin-transform-es2015-template-literals.
Prior to this change, when `babel-plugin-transform-es2015-template-literals`
was present in the `plugins` array its `TaggedTemplateExpression` visitor was
always ran prior to this plugins `Identifier` plugin (because the
`TaggedTemplateExpression` is "seen" before the `Identifier`).

Since the `babel-plugin-transform-es2015-template-literals` plugin then transforms
the tagged template literal into a custom transpiled version (for use in runtime)
the `babel-plugin-htmlbars-inline-precompile` plugin never functioned properly.

The refactor also seems to simplify (IMO) and make this a bit easier to reason about.