Add ability to pass static options to transpilation.
This allows consumers to specify additional options during compilation,
for example you might need to ensure that the compiled template has a
moduleName set (e.g. if you are testing an AST plugin that needs to know
the compiled templates module name).

Note: this functionality is only possible with the `CallExpression`
version of `hbs`. This is because there is no way to pass additional
arguments to a `TaggedTemplateExpression`.