Ensure `*/` included in a template does break surrounding JS.
Prior to this a template like:

```js
hbs`some content */`
```

Would emit invalid JS:

```js
Ember.HTMLBars.template(
  /*
    some content */
  */
  {
    // some stuff emitted by actual compilation
  }
)
```

The comment showing the original template was added as a developer aide
(makes looking at files with many inline templates such as tests much
easier), but without escaping `*/` we risk the template content itself
breaking the JS parsing within the file.

This commit ensures that `*/` is escaped to `*\/` which avoids the
issue.