Parse the precompilation result instead of using `babel.template`.
When using `@babel/template` (which is what we were using before) it
replaces the placeholders as identifiers. This results in a
fundamentally invalid AST (we would be having an identifier with a
`.name` property that was a JSON object) and that discrepancy caused
issues for other babel plugins that were trying to sanitize identifier
names (e.g. `@babel/plugin-transform-unicode-escapes` which was trying
to replace any unicode characters in identifiers).