Add `moduleOverrides` options
Adds a global `moduleOverrides` option, which can be used to configure
the import paths that are used for various imports in this transform.
This is meant to be used by Glimmer.js.

Also fixes the import path used for `templateOnly`, which is
meant to be the `default` export of `@ember/component/template-only`