Default `ensureModuleApiPolyfill` to true
When we released the refactor to transform to module names instead of
the API polyfill, sometimes those modules would transform correctly,
and sometimes they wouldn't. There appears to be a race condition
between `babel-plugin-ember-modules-api-polyfill`, this plugin, and the
AMD module transform.

This PR updates the default for `ensureModuleApiPolyfill` to true to
make sure we always transpile modules, and do not regress behavior.